requirements.txt
pandas==1.3.4
requests
psycopg2-binary
python-dotenv